refactor(cli): extract share logic to share-command.ts, address review feedback
Address review feedback from Carlos and CodeRabbit:

- Extract share handler from nemoclaw.ts into src/lib/share-command.ts
  following the existing *-command.ts pattern (gateway-token-command.ts,
  debug-command.ts). Thin dispatch in nemoclaw.ts injects dependencies.
- Probe fusermount3 before fusermount for FUSE 3 compat (CodeRabbit)
- Distinguish default vs custom --from sandboxes in sftp error (CodeRabbit)
- Use mountpoint -q on Linux with mount-parsing fallback (CodeRabbit)
- Assert non-zero exit in share-action recognition test (CodeRabbit)
